+++ Tyuiu.GrigorjanAM.Sprint3.Task0.V11/Program.cs
@@ -30,7 +30,7 @@ namespace Tyuiu.GrigorjanAM.Sprint3.Task0.V11
             Console.WriteLine("* ИСХОДНЫЕ ДАННЫЕ:                                                        *");
             Console.WriteLine("***************************************************************************");
 
-            double value = 5;
+            int value = 5;
             int startValue = 1;
             int stopValue = 10;
